requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aioboto3=11.2.0=pypi_0
aiobotocore=2.5.0=pypi_0
aiohttp=3.8.4=pypi_0
aioitertools=0.11.0=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.5.0=py311hca03da5_0
appdirs=1.4.4=pyhd3eb1b0_0
appnope=0.1.2=py311hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h80987f9_0
arxiv=1.4.7=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=pypi_0
boto3=1.26.76=pypi_0
botocore=1.29.76=pypi_0
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py311h80987f9_1002
bs4=0.0.1=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.4.26=hca03da5_0
cachetools=5.3.1=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.1=py311h80987f9_3
chardet=5.1.0=pypi_0
charset-normalizer=3.1.0=pypi_0
ci-info=0.3.0=pypi_0
click=8.1.3=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.2=py311hca03da5_0
configobj=5.0.8=pypi_0
configparser=5.3.0=pypi_0
contourpy=1.0.5=py311h48ca7d4_0
cryptography=39.0.1=py311h834c97f_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses-json=0.5.7=pypi_0
debugpy=1.5.1=py311h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
deeplake=3.6.11=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.6=pypi_0
entrypoints=0.4=py311hca03da5_0
et-xmlfile=1.1.0=pypi_0
etelemetry=0.3.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
faiss-cpu=1.7.4=pypi_0
feedparser=6.0.10=pypi_0
filelock=3.12.0=pypi_0
fitz=0.0.1.dev2=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=pypi_0
fsspec=2023.5.0=pypi_0
future=0.18.3=pypi_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h80987f9_3
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
glib=2.69.1=h98b2900_1
google-search-results=2.4.2=pypi_0
gptcache=0.1.34=pypi_0
gst-plugins-base=1.14.1=hf0a386a_0
gstreamer=1.14.1=he09cfb7_0
httplib2=0.22.0=pypi_0
huggingface-hub=0.14.1=pypi_0
humanfriendly=10.0=pypi_0
humbug=0.3.2=pypi_0
icu=68.1=hc377ac9_0
idna=3.4=py311hca03da5_0
importlib-metadata=6.8.0=pypi_0
ipykernel=6.19.2=py311h37496c9_0
ipython=8.12.0=py311hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py311hca03da5_0
isodate=0.6.1=pypi_0
jedi=0.18.1=py311hca03da5_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=1.0.1=pypi_0
joblib=1.1.1=py311hca03da5_0
jpeg=9e=h80987f9_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py311hca03da5_0
jupyter=1.0.0=py311hca03da5_8
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=5.3.0=py311hca03da5_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
jupyterlab_widgets=3.0.5=py311hca03da5_0
kiwisolver=1.4.4=py311h313beb8_0
krb5=1.19.2=h3b8d789_0
langchain=0.0.235=pypi_0
langchainplus-sdk=0.0.16=pypi_0
langsmith=0.0.7=pypi_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libclang=12.0.0=default_hc321e17_4
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20210910=h1a28f6b_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libllvm12=12.0.0=h12f7ac0_4
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libpq=12.9=h65cfe13_3
libsodium=1.0.18=h1a28f6b_0
libtiff=4.5.0=h313beb8_2
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
llvm-openmp=14.0.6=hc6e5704_0
looseversion=1.2.0=pypi_0
lxml=4.9.2=pypi_0
lz4-c=1.9.4=h313beb8_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py311h80987f9_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matplotlib=3.7.1=py311hca03da5_1
matplotlib-base=3.7.1=py311h7aedaa7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mdurl=0.1.2=pypi_0
mistune=0.8.4=py311h80987f9_1000
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
munkres=1.1.4=py_0
mypy-extensions=1.0.0=pypi_0
nb_conda_kernels=2.3.1=py311hca03da5_0
nbclassic=0.5.5=py311hca03da5_0
nbconvert=5.5.0=py_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=3.1=pypi_0
nibabel=5.1.0=pypi_0
nipype=1.8.6=pypi_0
nltk=3.8.1=pypi_0
notebook=6.5.4=py311hca03da5_0
notebook-shim=0.2.2=py311hca03da5_0
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numcodecs=0.11.0=pypi_0
numexpr=2.8.4=py311hdeb1df9_0
numpy=1.24.3=py311hb57d4eb_0
numpy-base=1.24.3=py311h1d85a46_0
onnxruntime=1.15.1=pypi_0
openai=0.27.8=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1t=h1a28f6b_0
packaging=23.1=pypi_0
pandas=1.5.3=py311h6956b77_0
pandoc=2.12=hca03da5_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathlib=1.0.1=pypi_0
pathos=0.3.0=pypi_0
pcre=8.45=hc377ac9_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py311h313beb8_0
pip=23.0.1=py311hca03da5_0
platformdirs=2.5.2=py311hca03da5_0
ply=3.11=py311hca03da5_0
pooch=1.4.0=pyhd3eb1b0_0
pox=0.3.2=pypi_0
ppft=1.7.6.6=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py311hca03da5_0
prompt_toolkit=3.0.36=hd3eb1b0_0
protobuf=4.23.3=pypi_0
prov=2.0.0=pypi_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=12.0.1=pypi_0
pyarxiv=1.0.3.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.7=pypi_0
pydeck=0.8.1b0=pypi_0
pydot=1.4.2=pypi_0
pygments=2.15.1=pypi_0
pyjwt=2.7.0=pypi_0
pympler=1.0.1=pypi_0
pymupdf=1.22.3=pypi_0
pyopenssl=23.0.0=py311hca03da5_0
pyparsing=3.0.9=py311hca03da5_0
pypdf2=3.0.1=pypi_0
pyqt=5.15.7=py311h313beb8_0
pyqt5-sip=12.11.0=py311h313beb8_0
pyrsistent=0.18.0=py311h80987f9_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.3=hc0d8a6c_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
pytz=2022.7=py311hca03da5_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyxnat=1.5=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=py311h313beb8_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_4
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
qtwebkit=5.212=h0f11f3c_4
rdflib=6.3.2=pypi_0
readline=8.2=h1a28f6b_0
regex=2023.5.5=pypi_0
requests=2.29.0=py311hca03da5_0
rich=13.4.2=pypi_0
s3transfer=0.6.1=pypi_0
scikit-learn=1.2.2=py311h313beb8_0
scipy=1.10.1=py311hc76d9b0_0
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=66.0.0=py311hca03da5_0
sgmllib3k=1.0.0=pypi_0
simplejson=3.19.1=pypi_0
sip=6.6.2=py311h313beb8_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py311hca03da5_1
soupsieve=2.4.1=pypi_0
sqlalchemy=1.4.48=pypi_0
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.24.1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.2=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=py311hca03da5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.2=py311h80987f9_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py311hca03da5_0
traits=6.3.2=pypi_0
transformers=4.29.2=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.8.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2023.3=pypi_0
tzlocal=4.3.1=pypi_0
urllib3=1.26.15=py311hca03da5_0
validators=0.20.0=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311hca03da5_1
websocket-client=0.58.0=py311hca03da5_4
wheel=0.38.4=py311hca03da5_0
widgetsnbextension=4.0.5=py311hca03da5_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h80987f9_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=hc377ac9_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
